Fix eval metrics type not json serializable.

+++ src/prime_rl/orchestrator/eval_utils.py
@@ -120,9 +120,9 @@ async def evaluate_env(
 
     # Log statistics to monitor
     eval_metrics = {
-        f"avg@{rollouts_per_example}": results_df.reward.mean(),
-        "no_response/mean": results_df.no_response.mean(),
-        "no_response/count": results_df.no_response.sum(),
+        f"avg@{rollouts_per_example}": float(results_df.reward.mean()),
+        "no_response/mean": float(results_df.no_response.mean()),
+        "no_response/count": int(results_df.no_response.sum()),
         "completion_len/mean": results_df.completion_len.mean().item(),
         "completion_len/max": results_df.completion_len.max().item(),
         "completion_len/min": results_df.completion_len.min().item(),
